+++ crates/config/src/schema/renderers/json_schema.rs
@@ -17,7 +17,7 @@ pub struct JsonSchemaRenderer {
 }
 
 fn clean_comment(comment: String) -> String {
-    comment.trim().replace("\n", " ")
+    comment.trim().replace('\n', " ")
 }
 
 impl JsonSchemaRenderer {