+++ crates/macros/src/config/setting.rs
@@ -99,7 +99,7 @@ impl<'l> Setting<'l> {
         } else if let Some(serde) = &self.serde_args.rename {
             serde.to_owned()
         } else if let Some(format) = casing_format {
-            format_case(format, &self.name.to_string())
+            format_case(format, &self.name.to_string(), false)
         } else {
             self.name.to_string()
         }